Refactor: Integrate v0.10.0 feature tests into main test suite
- Created `tests/test_simulation_config.py` for config tests.
- Updated `tests/test_simulation.py` with metrics and system event tests.
- Updated `tests/test_models.py` with Persona test.
- Removed temporary `tests/test_v0_10_0_changes.py`.
- Verified all tests pass with 100% coverage.
